# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "freddiemokwena949@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/order": {
            "get": {
                "summary": "Get all orders",
                "operationId": "order",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "summary": "Post a order",
                "operationId": "order",
                "parameters": [
                    {
                        "name": "Discount",
                        "in": "path",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "Total_Amount",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "Latitude",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "Longitude",
                        "in": "path",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/api/order/{id}": {
            "get": {
                "summary": "Get a order",
                "operationId": "order",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "put": {
                "summary": "Edit a order",
                "operationId": "order",
                "parameters": [
                    {
                        "name": "Discount",
                        "in": "path",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "Total_Amount",
                        "in": "path",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "Latitude",
                        "in": "path",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "Longitude",
                        "in": "path",
                        "required": false,
                        "type": "number"
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a order",
                "operationId": "order",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}